requirements.txt
aiogram>=3.2.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
aiohttp>=3.9.3
pydantic>=2.6.4
python-srt>=1.1.3
loguru>=0.7.2
